Start cleaning up and splitting the code
