+++ docs/components/Preview.js
@@ -1,4 +1,5 @@
 import CodeMirror from "codemirror"
+import CodeTabs from "../utils/tabs.js"
 
 function format(node, level) {
   const indentBefore = new Array(level++ + 1).join("  ")
@@ -20,27 +21,6 @@ function format(node, level) {
   return node
 }
 
-function initTabs() {
-  const tabs = document.querySelectorAll(".vueds-tab")
-  if (tabs) {
-    tabs.forEach(function(element) {
-      element.addEventListener("click", event => {
-        event.preventDefault()
-        document.querySelector(".vueds-tab--active").classList.remove("vueds-tab--active")
-        element.classList.add("vueds-tab--active")
-        document.querySelector(".vueds-hidden").classList.remove("vueds-hidden")
-        if (event.target.classList.contains("html")) {
-          const container = document.querySelector("article div[class^='rsg--tab']")
-          container.querySelectorAll("div")[0].classList.add("vueds-hidden")
-        } else {
-          const container = document.querySelector("article div[class^='rsg--tab']")
-          document.querySelector(".vueds-html").classList.add("vueds-hidden")
-        }
-      })
-    })
-  }
-}
-
 export default previewComponent => {
   // https://vuejs.org/v2/guide/render-function.html
   return {
@@ -48,18 +28,8 @@ export default previewComponent => {
       return createElement(previewComponent)
     },
     mounted() {
-      const oldElem = document.querySelector(".vueds-html")
-      const oldTabs = document.querySelector(".vueds-tabs")
-      if (oldElem || oldTabs) {
-        oldElem.parentNode.removeChild(oldElem)
-        oldTabs.parentNode.removeChild(oldTabs)
-      }
-
-      const tabs = document.createElement("div")
-      tabs.className = "vueds-tabs"
-      tabs.innerHTML =
-        "<button class='vueds-tab vue vueds-tab--active'>VUE</button><button class='vueds-tab html'>HTML</button>"
-
+      CodeTabs.clean()
+      const tabs = CodeTabs.create()
       const strDiv = this.$el.innerHTML.replace(/<!---->/g, "").replace(/data-v-\w*=""/g, "")
       const div = document.createElement("div")
       div.innerHTML =
@@ -100,7 +70,7 @@ export default previewComponent => {
             }
           )
 
-          initTabs()
+          CodeTabs.init()
         }, 150)
       }
     },
